test: migrate tests
